refactor backtest
